FinSync-main/attached_assets/requirements_1755681198277.txt
langgraph
python-dotenv
google-generativeai
pandas
openpyxl
